Fixed ground physics and inserted "player" object into Ground Template prefab AK collision Enter drawer
Just fixed two minor things as listed above. Collisions were happening randomly between the walls and the wood causing sounds events to trigger. Also, with the floor's physic set to "walls" the ball was bouncing higher with each bounce. I set it to none and it works. We could probably create physics for each floor type. Also made the concrete texture smaller so the look matches the sound better.